adds jailkit and some service version facts

+++ manifests/jailkit/setup.pp
@@ -12,39 +12,41 @@ class isp3node::jailkit::setup(
   String $checksum_type,
   String $tmpfolder,
 ) {
-  package{$build_packages:
-    ensure => latest,
-    tag    => ['build-req'],
-  }
-  -> archive { $file:
-    ensure        => present,
-    path          => "/tmp/${file}",
-    extract       => true,
-    extract_path  => '/tmp',
-    source        => "${source}/${file}",
-    checksum      => $checksum,
-    checksum_type => $checksum_type,
-    creates       => "/tmp/${tmpfolder}",
-    cleanup       => true,
-  }
-  -> file{"/tmp/${tmpfolder}/debian/comapat":
-    ensure  => file,
-    content => '5',
-  }
-  -> exec{'jailkit-build':
-    command => 'rules binary && mv /tmp/jailkit_*.deb /tmp/jailkit.deb',
-    path    => "${facts['path']}:/tmp/${tmpfolder}/debian",
-    cwd     => "/tmp/${tmpfolder}",
-    unless  => 'ls /tmp/jailkit.deb',
-  }
-  -> package{'jailkit':
-    ensure   => installed,
-    provider => 'dpkg',
-    source   => '/tmp/jailkit.deb',
-  }
-  -> exec{'jailkit-cleanup':
-    command => 'rm -rf /tmp/jailkit*',
-    onlyif  => 'ls /tmp/jailkit*',
-    path    => $facts['path'],
+  unless($facts['isp3node']['jailkit']['installed']){
+    package{$build_packages:
+      ensure => latest,
+      tag    => ['build-req'],
+    }
+    -> archive { $file:
+      ensure        => present,
+      path          => "/tmp/${file}",
+      extract       => true,
+      extract_path  => '/tmp',
+      source        => "${source}/${file}",
+      checksum      => $checksum,
+      checksum_type => $checksum_type,
+      creates       => "/tmp/${tmpfolder}",
+      cleanup       => true,
+    }
+    -> file{"/tmp/${tmpfolder}/debian/comapat":
+      ensure  => file,
+      content => '5',
+    }
+    -> exec{'jailkit-build':
+      command => 'rules binary && mv /tmp/jailkit_*.deb /tmp/jailkit.deb',
+      path    => "${facts['path']}:/tmp/${tmpfolder}/debian",
+      cwd     => "/tmp/${tmpfolder}",
+      unless  => 'ls /tmp/jailkit.deb',
+    }
+    -> package{'jailkit':
+      ensure   => installed,
+      provider => 'dpkg',
+      source   => '/tmp/jailkit.deb',
+    }
+    -> exec{'jailkit-cleanup':
+      command => 'rm -rf /tmp/jailkit*',
+      onlyif  => 'ls /tmp/jailkit*',
+      path    => $facts['path'],
+    }
   }
 }
